Add files via upload
This script will create a Jira issue (incident) for each Nagios incident.
The flow direction is from Nagios to Jira where This perl script would be required to be added in Nagios XI configuration as global event handler for host and services. This is early version 0.02 and required refinement to fulfill other condition such as duplicate avoidance and creating a unique title for each ticket.